Fixed a crash on report_crash() in cocoa.__init__.
The thing was that sometimes, we end up with a traceback that references pyc files directly and then traceback.extract_tb(), when it tries to fetch lines of code fetches binary data and crashes when trying to decode it. I implemented a safe_format_exception() function that falls back to a traceback formatting function that doesn't do any line fetching.